Relocate CLI IO and Template from `grails.cli.commands` to `grails.cli.core`
