+++ grace-cli/src/main/groovy/grails/cli/core/template/TemplateException.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package grails.cli.commands.template
+package grails.cli.core.template
 
 import groovy.transform.InheritConstructors
