+++ grace-cli/src/main/groovy/grails/cli/core/template/TemplateRenderer.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package grails.cli.commands.template
+package grails.cli.core.template
 
 import groovy.transform.CompileDynamic
 import org.springframework.core.io.Resource
